% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{apple_health_export_dir}
\alias{apple_health_export_dir}
\title{Apple Health export directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
apple_health_export_dir
}
\description{
Apple Health export directory
}
\keyword{datasets}
