% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_from_export}
\alias{get_data_from_export}
\title{Get data from export file}
\usage{
get_data_from_export(x)
}
\arguments{
\item{x}{Data file}
}
\description{
Get data from export file downloaded from Apple Watch.
}
