% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heart_rate.R
\name{get_heart_rate}
\alias{get_heart_rate}
\title{Get heart rate from data records}
\usage{
get_heart_rate(x)
}
\arguments{
\item{x}{Data file}
}
\description{
Get data from export file downloaded from Apple Watch.
}
