% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{heart_rate_file}
\alias{heart_rate_file}
\title{Heart rate file}
\format{
An object of class \code{character} of length 1.
}
\usage{
heart_rate_file
}
\description{
Heart rate file
}
\keyword{datasets}
