% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{watch_export_file}
\alias{watch_export_file}
\title{Apple Watch export file}
\format{
An object of class \code{character} of length 1.
}
\usage{
watch_export_file
}
\description{
Apple Watch export file
}
\keyword{datasets}
